feat: oracle redo parse update and delete commit

+++ src/main/java/com/example/redo/model/RedoRecord.java
@@ -9,7 +9,7 @@ public record RedoRecord(
         int length,
         int headerLength,
         int vld,
-        int scn,
+        long scn,
         int subScn,
         List<RedoChange> changes
 ) {
